chore: move the collaborator api to pro version
