+++ packages/global/core/app/collaborator.d.ts
@@ -1,0 +1,12 @@
+import { PermissionValueType } from '../../support/permission/type';
+
+export type AddAppCollaboratorRequest = {
+  appId: string;
+  tmbIds: string[];
+  permission?: PermissionValueType;
+};
+
+export type AppCollaboratorDeleteParams = {
+  appId: string;
+  tmbId: string;
+};
