+++ projects/app/src/web/core/app/collaborator.ts
@@ -1,16 +1,18 @@
-import { AddAppCollaboratorRequest } from '@/pages/api/core/app/collaborator/add';
-import { AppCollaboratorDeleteParams } from '@/pages/api/core/app/collaborator/delete';
+import {
+  AddAppCollaboratorRequest,
+  AppCollaboratorDeleteParams
+} from '@fastgpt/global/core/app/collaborator';
 import { DELETE, GET, POST } from '@/web/common/api/request';
 import { AppCollaboratorType } from '@fastgpt/global/core/app/type';
 
 export const getCollaboratorList = (appId: string) =>
-  GET<AppCollaboratorType[]>('/core/app/collaborator/list', { appId });
+  GET<AppCollaboratorType[]>('/proApi/core/app/collaborator/list', { appId });
 
 export const addAppCollaborators = ({ ...parms }: AddAppCollaboratorRequest) =>
-  POST('/core/app/collaborator/add', { ...parms });
+  POST('/proApi/core/app/collaborator/add', { ...parms });
 
 export const updateAppCollaborators = ({ ...parms }: AddAppCollaboratorRequest) =>
-  POST('/core/app/collaborator/add', { ...parms }); // the same api with add
+  POST('/proApi/core/app/collaborator/add', { ...parms }); // the same api with add
 
 export const deleteAppCollaborators = ({ ...params }: AppCollaboratorDeleteParams) =>
-  DELETE('/core/app/collaborator/delete', { ...params });
+  DELETE('/proApi/core/app/collaborator/delete', { ...params });
